Escape Node Names in Messages
The \mess macro uses the text of the
message between nodes to create a node for later annotations. If the
text contains a parentheses or other special TikZ character, the
document will fail to compile.

This update uses the xstring package to replace all special characters
in the text of the message with "*", so the node is still created but
with a legal name.